pyproject.toml
[tool.poetry]
name = "mediathequeroubaix"
version = "0.1.0"
description = "Client for the library of Roubaix (Médiathèque Roubaix)"
authors = [
    "Thomas Queste <tom@tomsquest.com>",
]
license = "MIT"
readme = "README.md"

documentation = "https://github.com/tomsquest/mediathequeroubaix.py"
homepage = "https://tomsquest.github.io/mediathequeroubaix.py"
repository = "https://github.com/tomsquest/mediathequeroubaix.py"

classifiers = [
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Typing :: Typed",
]

packages = [
    { include = "mediathequeroubaix", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.7.1, <4.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-eradicate = "*"
flake8-logging-format = "*"
isort = "*"
# Force version of mypy to the one of Results to avoid weird errors
mypy = "^0.950"
pep8-naming = "*"
pre-commit = "*"
pytest = "*"
pytest-github-actions-annotate-failures = "*"
pytest-cov = "*"
python-kacl = "*"
pyupgrade = "*"
tryceratops = "*"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.28.11.2"
returns = "^0.19.0"
requests-mock = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.black]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = """\
    --cov mediathequeroubaix \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.report]
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover'
]

[tool.mypy]
disallow_any_unimported = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true
strict = true
plugins = ['returns.contrib.mypy.returns_plugin', 'pydantic.mypy']
